feat(web): Housekeeping becomes Insights, and /ui/ops points at it
The page moves to src/web/insights.rs with the ten row structs only it
uses; ui.rs drops from 10,880 to 10,490 lines. The POST /ui/ops/...
actions stay where they are — artifact_changed and ReturnTo serve
handlers all over ui.rs, and dragging them across a file boundary would
churn unrelated code for no gain. Their paths are unchanged either way.

source_rows stays too, widened rather than copied: its own comment says
a copy in each is how Ops and the detail pane come to disagree about
what a merge was made of.

/ui/ops answers 303 rather than 404 — it is in bookmarks and in at
least one runbook. The redirect takes an Identity like every other /ui
route: every_ui_route_requires_a_session caught it when it did not, and
a redirect that answers before the session is checked tells an
anonymous caller which paths exist.
