+++ src/web/templates/capture.html
@@ -132,7 +132,7 @@
    rather than about what is in it. Neither belongs in the top row, which stays
    three destinations wide. #}
 <p class="quiet-links">
-  <a class="quiet-link" href="/ui/ops">Housekeeping</a>
+  <a class="quiet-link" href="/ui/insights">Insights</a>
   <a class="quiet-link" href="/ui/settings">Settings</a>
 </p>
 </div>
