+++ src/web/templates/insights.html
