+++ src/web/templates/layout.html
@@ -73,6 +73,11 @@
       <a href="/ui/judge">Judge{% if n > &0 %}
         <span class="badge badge-accent">{{ n }}</span>{% endif %}</a>
       {% endif %}
+      {# What is true about this installation, and what needs a person: the
+         maintenance Housekeeping always held, plus the surfaces that used to
+         sit on Capture. A destination rather than a quiet link at the bottom
+         of a page, because it is now the only other place there is. #}
+      <a href="/ui/insights">Insights</a>
       <span class="spacer"></span>
       {# The light palette has been in the stylesheet since the port and nobody
          has ever seen it, because it activated only on prefers-color-scheme.
