+++ src/web/ui.rs
@@ -224,39 +224,6 @@ pub struct SeenTogether {
     pub cross_corpus: bool,
 }
 
-/// Work that hit something and is waiting to try again by itself.
-pub struct RetryingRow {
-    pub stage: String,
-    pub target_id: String,
-    pub attempts: i64,
-    pub due: String,
-    pub last_error: String,
-}
-
-/// A parked capture, with enough of the corpus it resembles to decide without
-/// opening both.
-pub struct ParkedRow {
-    pub id: String,
-    pub title: String,
-    pub bytes: usize,
-    pub other_id: String,
-    pub other_title: String,
-    pub percent: i64,
-}
-
-/// An artifact the sweep hid, with the one it lost to.
-pub struct SupersededRow {
-    pub id: String,
-    pub title: String,
-    /// When it was written and how it opens. Two artifacts can carry the same
-    /// title — a merge of two documents that named a section identically
-    /// produces exactly that — and a table of them is unreadable without
-    /// something that differs between the rows.
-    pub subtitle: String,
-    pub winner_id: String,
-    pub winner_title: String,
-}
-
 /// A pair waiting on a person.
 pub struct PairRow {
     pub id: i64,
@@ -295,19 +262,6 @@ pub struct PairRow {
     /// still resolvable, just with nothing recommended.
     pub keeps_a: bool,
     pub keeps_b: bool,
-}
-
-/// An artifact flagged stale with no specific replacement.
-pub struct DeprecatedRow {
-    pub id: String,
-    pub title: String,
-}
-
-/// An active artifact nobody has confirmed or retrieved in a while.
-pub struct StaleRow {
-    pub id: String,
-    pub title: String,
-    pub last_verified: String,
 }
 
 pub struct TokenRow {
@@ -366,7 +320,7 @@ pub fn fmt_duration(secs: i64) -> String {
 /// An identifier with no wording yet returns unchanged rather than blank: a
 /// stage added later must show up as *something*. `every_stage_the_queue_can_run_has_a_word_for_it`
 /// is what makes sure that fallback stays theoretical.
-fn sweep_label(stage: &str) -> &str {
+pub(crate) fn sweep_label(stage: &str) -> &str {
     match stage {
         "synthesize" => "Writing artifacts",
         "enrich" => "Enriching",
@@ -793,29 +747,8 @@ const SWEEP_WORDS: &[(&str, &str, &str)] = &[
     ("pursuit", "pursuits", "pursuits opened"),
 ];
 
-/// One phrase of the last day: "412 links forgotten".
-struct SweepCount {
-    n: i64,
-    what: String,
-}
-
-/// One recorded run, as the history renders it.
-struct SweepRunRow {
-    when: String,
-    /// The stage in words. The identifier it was worded from is on the cell as
-    /// a `title`, because the log and the config still call it that and a
-    /// reader who greps for `arm_dedupe` should find it here too.
-    stage: String,
-    stage_id: String,
-    /// Empty unless it failed, in which case it is why.
-    error: String,
-    took: String,
-    /// The counts, already worded. Empty for a run that did nothing.
-    counts: Vec<SweepCount>,
-}
-
 /// Add up one run's `detail` into `totals`, keyed by the words it earns.
-fn tally_sweep(stage: &str, detail: &str, totals: &mut Vec<(String, i64)>) {
+pub(crate) fn tally_sweep(stage: &str, detail: &str, totals: &mut Vec<(String, i64)>) {
     let Ok(v) = serde_json::from_str::<serde_json::Value>(detail) else {
         return;
     };
@@ -838,67 +771,6 @@ fn tally_sweep(stage: &str, detail: &str, totals: &mut Vec<(String, i64)>) {
 }
 
 #[derive(Template)]
-#[template(path = "ops.html")]
-struct OpsTemplate {
-    /// Waiting judgements for the nav. See `state::judge_pending`.
-    judge_pending: Option<i64>,
-    /// Whether the ask door is open. See `state::ask_enabled`.
-    ask_enabled: bool,
-    job_counts: Vec<(String, i64)>,
-    oldest_pending_secs: Option<i64>,
-    artifact_count: i64,
-    vector_count: u64,
-    retrying: Vec<RetryingRow>,
-    parked: Vec<ParkedRow>,
-    superseded: Vec<SupersededRow>,
-    /// Artifacts the dedupe pass wrote out of several others, with what they
-    /// were written from and an undo.
-    merged: Vec<MergedRow>,
-    /// The list is capped; there are rows this page is not showing. Said out
-    /// loud, because a table that stops without saying so reads as a table of
-    /// everything there is.
-    more_merged: bool,
-    more_superseded: bool,
-    /// `TABLE_CAP`, so the line that says how many rows are showing says the
-    /// number the code actually truncated to. Written out twice in the
-    /// template, it drifted from the constant the first time either moved.
-    table_cap: i64,
-    deprecated: Vec<DeprecatedRow>,
-    stale: Vec<StaleRow>,
-    /// `None` when nothing is being learned, which renders nothing at all: a
-    /// count of links on a base that records no searches is a line about a
-    /// feature that is switched off.
-    links: Option<crate::store::links::LinkCounts>,
-    /// Artifacts written from pursuits, newest first, each one click from
-    /// deprecated.
-    generated: Vec<GeneratedRow>,
-    /// Recent pursuits, only when the feature is on. A count and not a table:
-    /// a pursuit that ended unsatisfied is a hole in the base and belongs on
-    /// the one list of those, not on a second list of its own; one that ended
-    /// satisfied needs nobody; and one that was written up is in `generated`
-    /// above.
-    pursuit_enabled: bool,
-    pursuit_recent: usize,
-    pursuit_unsatisfied: usize,
-    /// What the sweeps did in the last twenty-four hours, added up. Not "last
-    /// night": units that reschedule themselves on their own periods do not
-    /// line up into one cycle, and there is no cycle identity to group them by.
-    last_day: Vec<SweepCount>,
-    /// Runs in the last day that failed. Said separately, because a summary of
-    /// what got done cannot report what did not.
-    last_day_failures: usize,
-    /// The runs themselves, newest first. What a single overwritten summary
-    /// could never give: whether this started yesterday or has been going
-    /// wrong for a week.
-    sweep_history: Vec<SweepRunRow>,
-    /// Shown against clicked, by rung. Empty when the offer is switched off, or
-    /// when it has been on and never had anything to say — either way there is
-    /// no table, because a heading over no rows is a claim that something is
-    /// being measured when nothing is.
-    offer_rates: Vec<crate::store::pursuits::OfferRate>,
-}
-
-#[derive(Template)]
 #[template(path = "settings.html")]
 struct SettingsTemplate {
     /// Waiting judgements for the nav. See `state::judge_pending`.
@@ -916,27 +788,6 @@ struct SettingsTemplate {
     asks: Option<crate::store::asks::AskStats>,
 }
 
-/// One generated artifact on Ops.
-struct GeneratedRow {
-    id: String,
-    title: String,
-    subtitle: String,
-    cues: Vec<String>,
-    sources: Vec<SourceRow>,
-}
-
-struct MergedRow {
-    id: String,
-    title: String,
-    /// See `SupersededRow::subtitle`: what tells two rows with one title apart.
-    subtitle: String,
-    /// What it was written from, in the order the lineage stores them.
-    sources: Vec<SourceRow>,
-    /// True when a source has been deleted since, so the artifact claims less
-    /// provenance than its text carries.
-    orphaned: bool,
-}
-
 pub struct SourceRow {
     pub id: String,
     pub title: String,
@@ -950,7 +801,7 @@ pub struct SourceRow {
 
 /// When an artifact was written and how it opens, for a table where the title
 /// alone may not be unique.
-fn row_subtitle(c: &crate::store::artifacts::Chunk) -> String {
+pub(crate) fn row_subtitle(c: &crate::store::artifacts::Chunk) -> String {
     format!(
         "{} · {}",
         fmt_time(c.created_at),
@@ -963,7 +814,7 @@ fn row_subtitle(c: &crate::store::artifacts::Chunk) -> String {
 /// identical (same self-guard, same tolerance for deleted sources, same
 /// corpus fallback), and a copy in each is how they come to disagree about
 /// what a merge was made of.
-async fn source_rows(
+pub(crate) async fn source_rows(
     store: &crate::store::Store,
     merged_id: &str,
     roots: &[String],
@@ -2278,7 +2129,7 @@ async fn delete_artifact_ui(
     // where deleting one leaves you.
     Ok(match corpus_id {
         Some(cid) => Redirect::to(&format!("/ui/corpora/{cid}")).into_response(),
-        None => Redirect::to("/ui/ops").into_response(),
+        None => Redirect::to("/ui/insights").into_response(),
     })
 }
 
@@ -2595,278 +2446,6 @@ async fn settings(State(st): State<AppState>, _id: Identity) -> Result<Response>
     .into_response())
 }
 
-/// Rows of one housekeeping table before it says there are more.
-///
-/// These tables are read to answer "what happened to X", and the answer to
-/// that is a search for X rather than a scroll — so the cap is stated and the
-/// rest arrive as these are cleared, instead of growing a pager nobody would
-/// page through.
-const TABLE_CAP: i64 = 25;
-
-async fn ops(State(st): State<AppState>, _id: Identity) -> Result<Response> {
-    use sqlx::Row;
-
-    let artifact_count: i64 = sqlx::query("SELECT COUNT(*) AS n FROM artifacts")
-        .fetch_one(&st.core.store.pool)
-        .await?
-        .get("n");
-
-    // Not a queue of chores: work that hit something and is waiting to try
-    // again on its own. Nothing here needs a person.
-    let retrying: Vec<RetryingRow> = st
-        .core
-        .store
-        .retrying_jobs(50)
-        .await?
-        .into_iter()
-        .map(|j| RetryingRow {
-            stage: j.stage,
-            target_id: j.target_id,
-            attempts: j.attempts,
-            due: fmt_duration(j.next_attempt_secs),
-            last_error: j.last_error.unwrap_or_else(|| "—".into()),
-        })
-        .collect();
-
-    // A parked capture is the one corpus state no worker advances. It has to be
-    // shown here or it sits unprocessed with nothing saying why.
-    let mut parked = Vec::new();
-    for c in st.core.store.parked_corpora(50).await? {
-        let other_id = c.near_dupe_of.clone().unwrap_or_default();
-        let other_title = match st.core.store.get_corpus(&other_id).await {
-            Ok(o) => o.title_hint.unwrap_or_else(|| "untitled".into()),
-            Err(_) => "(deleted)".into(),
-        };
-        parked.push(ParkedRow {
-            percent: (c.near_dupe_score.unwrap_or(0.0) * 100.0).round() as i64,
-            bytes: c.raw_text.len(),
-            title: c.title_hint.clone().unwrap_or_else(|| "untitled".into()),
-            id: c.id,
-            other_id,
-            other_title,
-        });
-    }
-
-    let mut superseded = Vec::new();
-    // One past the cap, so the page can say it is capped rather than truncate
-    // in silence — a table that stops at 25 with nothing said reads as a table
-    // of everything there is.
-    for c in st.core.store.superseded_artifacts(TABLE_CAP + 1).await? {
-        let winner_id = c.superseded_by.clone().unwrap_or_default();
-        let winner_title = match st.core.store.get_artifact(&winner_id).await {
-            Ok(w) => title_of(&w),
-            Err(_) => "(deleted)".to_string(),
-        };
-        superseded.push(SupersededRow {
-            title: title_of(&c),
-            subtitle: row_subtitle(&c),
-            id: c.id,
-            winner_id,
-            winner_title,
-        });
-    }
-
-    let mut merged = Vec::new();
-    let merged_chunks = st.core.store.merged_artifacts(TABLE_CAP + 1).await?;
-    // One lineage call per page, not one per row: `roots_of` takes the batch.
-    let merged_ids: Vec<String> = merged_chunks.iter().map(|c| c.id.clone()).collect();
-    let roots = st
-        .core
-        .store
-        .roots_of(&merged_ids)
-        .await
-        .unwrap_or_default();
-    for c in merged_chunks {
-        let sources = source_rows(
-            &st.core.store,
-            &c.id,
-            roots.get(&c.id).map(Vec::as_slice).unwrap_or_default(),
-        )
-        .await;
-        merged.push(MergedRow {
-            orphaned: c.flags.iter().any(|f| f == "orphaned_source"),
-            title: title_of(&c),
-            subtitle: row_subtitle(&c),
-            id: c.id,
-            sources,
-        });
-    }
-
-    let more_merged = merged.len() > TABLE_CAP as usize;
-    merged.truncate(TABLE_CAP as usize);
-
-    let mut generated = Vec::new();
-    let gen_chunks = st.core.store.synthesized_artifacts(TABLE_CAP).await?;
-    let gen_ids: Vec<String> = gen_chunks.iter().map(|c| c.id.clone()).collect();
-    let gen_roots = st.core.store.roots_of(&gen_ids).await.unwrap_or_default();
-    for c in gen_chunks {
-        let sources = source_rows(
-            &st.core.store,
-            &c.id,
-            gen_roots.get(&c.id).map(Vec::as_slice).unwrap_or_default(),
-        )
-        .await;
-        generated.push(GeneratedRow {
-            title: title_of(&c),
-            subtitle: row_subtitle(&c),
-            cues: c.cues.clone(),
-            id: c.id,
-            sources,
-        });
-    }
-    let pursuit_enabled = st.core.learn.enabled;
-    let recent = match pursuit_enabled {
-        true => st.core.store.recent_pursuits(50).await?,
-        false => Vec::new(),
-    };
-    let pursuit_recent = recent.len();
-    // The ones the sentence below can honestly point at. `unsatisfied` is how a
-    // run of searches *ended*, and a capture that answers one afterwards leaves
-    // that word alone deliberately — coverage never rewrites what happened — so
-    // counting the state sent the operator to a gap list that had already
-    // dropped half of them.
-    let on_the_gap_list = match pursuit_enabled {
-        true => st
-            .core
-            .store
-            .open_pursuit_gap_ids(st.core.embedder.model())
-            .await
-            .unwrap_or_default(),
-        false => Default::default(),
-    };
-    let pursuit_unsatisfied = recent
-        .iter()
-        .filter(|p| p.state == "unsatisfied" && on_the_gap_list.contains(&p.id))
-        .count();
-    // What the memory did while nobody was looking. The last day as one
-    // sentence, and under it the runs themselves — which is the half a single
-    // overwritten summary could never give.
-    let day = st
-        .core
-        .store
-        .sweep_runs_since(crate::store::now() - 86_400, 500)
-        .await
-        .unwrap_or_default();
-    let last_day_failures = day.iter().filter(|r| r.outcome == "failed").count();
-    let mut totals: Vec<(String, i64)> = Vec::new();
-    for r in &day {
-        tally_sweep(&r.stage, &r.detail, &mut totals);
-    }
-    let last_day: Vec<SweepCount> = totals
-        .into_iter()
-        .map(|(what, n)| SweepCount { n, what })
-        .collect();
-    let sweep_history: Vec<SweepRunRow> = st
-        .core
-        .store
-        .sweep_history(TABLE_CAP)
-        .await
-        .unwrap_or_default()
-        .into_iter()
-        .map(|r| {
-            let mut counts = Vec::new();
-            tally_sweep(&r.stage, &r.detail, &mut counts);
-            SweepRunRow {
-                when: fmt_time(r.started_at),
-                error: match r.outcome == "failed" {
-                    true => serde_json::from_str::<serde_json::Value>(&r.detail)
-                        .ok()
-                        .and_then(|v| v.get("error").and_then(|e| e.as_str().map(String::from)))
-                        .unwrap_or_else(|| "it failed".into()),
-                    false => String::new(),
-                },
-                took: fmt_elapsed(r.ended_at - r.started_at),
-                stage: sweep_label(&r.stage).to_string(),
-                stage_id: r.stage,
-                counts: counts
-                    .into_iter()
-                    .map(|(what, n)| SweepCount { n, what })
-                    .collect(),
-            }
-        })
-        .collect();
-
-    let more_superseded = superseded.len() > TABLE_CAP as usize;
-    superseded.truncate(TABLE_CAP as usize);
-
-    let deprecated = st
-        .core
-        .store
-        .artifacts_by_status(crate::store::artifacts::ArtifactStatus::Deprecated, 50)
-        .await?
-        .into_iter()
-        .map(|c| DeprecatedRow {
-            title: title_of(&c),
-            id: c.id,
-        })
-        .collect();
-
-    // Read-only candidates: nothing here has been changed, only listed.
-    let stale = st
-        .core
-        .stale_candidates(50)
-        .await
-        .unwrap_or_else(|e| {
-            tracing::warn!(error = %e, "no stale candidates for ops");
-            vec![]
-        })
-        .into_iter()
-        .map(|r| StaleRow {
-            title: r.title.unwrap_or_else(|| markdown::snippet(&r.text, 60)),
-            id: r.artifact_id,
-            last_verified: r
-                .last_verified_at
-                .map(fmt_time)
-                .unwrap_or_else(|| "never".to_string()),
-        })
-        .collect();
-
-    Ok(HtmlTemplate(OpsTemplate {
-        judge_pending: crate::web::state::judge_pending(&st).await,
-        ask_enabled: crate::web::state::ask_enabled(&st),
-        retrying,
-        parked,
-        superseded,
-        merged,
-        more_merged,
-        more_superseded,
-        table_cap: TABLE_CAP,
-        deprecated,
-        stale,
-        job_counts: st.core.store.job_counts().await?,
-        oldest_pending_secs: st.core.store.oldest_pending_age().await?,
-        artifact_count,
-        // Qdrant being briefly unreachable must not blank the ops page, which
-        // is exactly where you look when something is wrong.
-        vector_count: st.core.vectors.count().await.unwrap_or(0),
-        links: match st.core.associating() {
-            true => Some(st.core.store.link_counts().await?),
-            false => None,
-        },
-        generated,
-        pursuit_enabled,
-        pursuit_recent,
-        pursuit_unsatisfied,
-        last_day,
-        last_day_failures,
-        sweep_history,
-        // The last month rather than the last day: a weekly pattern needs
-        // weeks, so a hit rate measured over a day would be a number nobody
-        // could act on. Read like `vector_count` — a failure here must not
-        // blank the page you open when something is wrong.
-        offer_rates: match st.core.recommends() {
-            true => st
-                .core
-                .store
-                .offer_rates(crate::store::now() - 30 * 86_400)
-                .await
-                .unwrap_or_default(),
-            false => Vec::new(),
-        },
-    })
-    .into_response())
-}
-
 /// Take a merge back: what it replaced returns, the merge is retired, and the
 /// pairs behind it are dismissed so the sweep does not simply redo it.
 async fn undo_merge_ui(
@@ -2875,7 +2454,7 @@ async fn undo_merge_ui(
     Path(aid): Path<String>,
 ) -> Result<Response> {
     crate::jobs::merge::undo(&st.core, &aid).await?;
-    Ok(Redirect::to("/ui/ops").into_response())
+    Ok(Redirect::to("/ui/insights").into_response())
 }
 
 /// Forget every captured search, every recorded question, and every situation.
@@ -2953,7 +2532,7 @@ async fn resolve_near_dupe_ui(
     Form(form): Form<ResolveForm>,
 ) -> Result<Response> {
     st.core.resolve_near_duplicate(&cid, form.action).await?;
-    Ok(Redirect::to("/ui/ops").into_response())
+    Ok(Redirect::to("/ui/insights").into_response())
 }
 
 /// Where a lifecycle button should land afterwards.
@@ -2975,7 +2554,7 @@ impl ReturnTo {
     fn path(&self) -> &str {
         match self.to.as_deref() {
             Some(p) if p.starts_with("/ui/") && !p.starts_with("/ui//") => p,
-            _ => "/ui/ops",
+            _ => "/ui/insights",
         }
     }
 }
@@ -3952,7 +3531,6 @@ pub fn ui_router() -> Router<AppState> {
         .route("/ui/ask/{id}/carried", post(ask_carried))
         .route("/ui/ask/{id}/keep", post(ask_keep))
         .route("/ui/gaps/{kind}/{id}/dismiss", post(gap_dismiss))
-        .route("/ui/ops", get(ops))
         // One name for the page. The nav word is Housekeeping and the route is
         // `/ui/ops`; a reader who types the word they were shown lands here.
         .route(
@@ -4949,6 +4527,38 @@ mod tests {
         app_with_cookie(core).await
     }
 
+    /// Housekeeping is Insights now. The old door stays a door: it is in
+    /// bookmarks, in the quiet link at the bottom of the page, and in at least
+    /// one runbook — a 404 there is a broken promise, not a tidy-up.
+    #[tokio::test]
+    async fn housekeeping_moved_to_insights_and_the_old_door_still_opens() {
+        let (app, cookie) = app_for(crate::core::test_support::test_core().await).await;
+
+        let html = get(&app, "/ui/insights", &cookie).await;
+        assert!(
+            html.contains("Housekeeping"),
+            "the maintenance section is there"
+        );
+
+        let res = app
+            .clone()
+            .oneshot(
+                Request::builder()
+                    .uri("/ui/ops")
+                    .header("cookie", &cookie)
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+        assert_eq!(res.status(), StatusCode::SEE_OTHER);
+        assert_eq!(
+            res.headers().get("location").unwrap(),
+            "/ui/insights",
+            "the old door points at the new one"
+        );
+    }
+
     #[tokio::test]
     async fn a_link_derived_pair_never_claims_a_similarity_once_the_judge_has_settled_it() {
         // The link judge files these with `detail = "link"` and a score of 0.0,
@@ -5834,7 +5444,7 @@ mod tests {
         assert_eq!(forgotten.opened, 0, "nobody took that one");
 
         let (app, cookie) = crate::web::test_support::app_with_cookie(core).await;
-        let page = get(&app, "/ui/ops", &cookie).await;
+        let page = get(&app, "/ui/insights", &cookie).await;
         assert!(page.contains("What was offered"), "no heading");
         assert!(page.contains("pattern"), "no rung");
         assert!(page.contains("forgotten"));
@@ -5846,7 +5456,7 @@ mod tests {
         // when nothing is.
         let core = crate::core::test_support::test_core().await;
         let (app, cookie) = crate::web::test_support::app_with_cookie(core).await;
-        let page = get(&app, "/ui/ops", &cookie).await;
+        let page = get(&app, "/ui/insights", &cookie).await;
         assert!(!page.contains("What was offered"));
     }
 
@@ -5900,7 +5510,7 @@ mod tests {
         // the page you open when something is wrong, which is the wrong place
         // for the screen that has to be visited often for the dataset to grow.
         let (app, cookie) = app_recording_searches(3).await;
-        for page in ["/ui/search", "/ui/capture", "/ui/ask", "/ui/ops"] {
+        for page in ["/ui/search", "/ui/capture", "/ui/ask", "/ui/insights"] {
             let html = flat(&get(&app, page, &cookie).await);
             assert!(
                 html.contains(r#"<a href="/ui/judge">Judge"#),
@@ -6440,7 +6050,7 @@ mod tests {
             ))
             .await
             .unwrap();
-        assert_eq!(res.headers().get("location").unwrap(), "/ui/ops");
+        assert_eq!(res.headers().get("location").unwrap(), "/ui/insights");
     }
 
     #[tokio::test]
@@ -6509,7 +6119,7 @@ mod tests {
                 .unwrap();
             assert_eq!(
                 res.headers().get("location").unwrap(),
-                "/ui/ops",
+                "/ui/insights",
                 "followed {hostile}"
             );
         }
@@ -7146,7 +6756,7 @@ mod tests {
             "/ui/queue",
             "/ui/corpora/abc",
             "/ui/ask",
-            "/ui/ops",
+            "/ui/insights",
         ] {
             // A plain GET is a browser loading a page, so a missing session
             // sends it to sign in rather than showing it JSON it cannot act
@@ -7600,7 +7210,7 @@ mod tests {
             .await
             .unwrap();
 
-        let html = get(&app, "/ui/ops", &cookie).await;
+        let html = get(&app, "/ui/insights", &cookie).await;
         assert!(
             html.contains("412 links forgotten"),
             "the last day did not add the runs up: {html}"
@@ -7618,7 +7228,7 @@ mod tests {
         let res = app
             .oneshot(
                 Request::builder()
-                    .uri("/ui/ops")
+                    .uri("/ui/insights")
                     .header("cookie", cookie)
                     .body(Body::empty())
                     .unwrap(),
@@ -7646,7 +7256,7 @@ mod tests {
             .await
             .unwrap();
 
-        let page = get_body(&app, &cookie, "/ui/ops").await;
+        let page = get_body(&app, &cookie, "/ui/insights").await;
         // One `bump_link` call between one pair is one row in `artifact_links`
         // — see `the_counts_say_how_many_links_there_are_and_how_many_are_named`
         // in store::links, which needs two calls between two different pairs
@@ -7670,7 +7280,7 @@ mod tests {
         let res = app
             .oneshot(
                 Request::builder()
-                    .uri("/ui/ops")
+                    .uri("/ui/insights")
                     .header("cookie", cookie)
                     .body(Body::empty())
                     .unwrap(),
@@ -7726,7 +7336,7 @@ mod tests {
             .clone()
             .oneshot(
                 Request::builder()
-                    .uri("/ui/ops")
+                    .uri("/ui/insights")
                     .header("cookie", &cookie)
                     .body(Body::empty())
                     .unwrap(),
@@ -7886,7 +7496,7 @@ mod tests {
     #[tokio::test]
     async fn the_counts_say_what_they_count() {
         let (app, cookie) = app_with_session().await;
-        let page = get_body(&app, &cookie, "/ui/ops").await;
+        let page = get_body(&app, &cookie, "/ui/insights").await;
         assert!(
             page.contains("jobs") || page.contains("No jobs queued"),
             "a job count must not read as an artifact count: {page}"
@@ -7902,7 +7512,7 @@ mod tests {
             .await
             .unwrap();
 
-        let page = get_body(&app, &cookie, "/ui/ops").await;
+        let page = get_body(&app, &cookie, "/ui/insights").await;
         assert!(!page.contains("Put it back"), "{page}");
         assert!(!page.contains("Undo merge"), "{page}");
         assert!(page.contains(">Undo<"), "{page}");
@@ -7917,7 +7527,7 @@ mod tests {
             .await
             .unwrap();
 
-        let page = get_body(&app, &cookie, "/ui/ops").await;
+        let page = get_body(&app, &cookie, "/ui/insights").await;
         assert!(
             page.contains("body of Windows Update-Typen"),
             "a row has to say which artifact it is, and two can share a title: {page}"
@@ -7960,7 +7570,7 @@ mod tests {
         assert!(settings.contains("API tokens"), "{settings}");
         assert!(settings.contains("Browser extension"), "{settings}");
 
-        let ops = get_body(&app, &cookie, "/ui/ops").await;
+        let ops = get_body(&app, &cookie, "/ui/insights").await;
         assert!(
             !ops.contains("API tokens"),
             "housekeeping is about the corpus: {ops}"
@@ -7972,7 +7582,7 @@ mod tests {
     async fn both_pages_are_reachable_from_capture() {
         let (app, cookie) = app_with_session().await;
         let page = get_body(&app, &cookie, "/ui/capture").await;
-        assert!(page.contains("/ui/ops"), "{page}");
+        assert!(page.contains("/ui/insights"), "{page}");
         assert!(page.contains("/ui/settings"), "{page}");
     }
 
@@ -9196,7 +8806,7 @@ mod tests {
             ids.push(p);
         }
 
-        let both = get_body(&app, &cookie, "/ui/ops").await;
+        let both = get_body(&app, &cookie, "/ui/insights").await;
         assert!(both.contains("2 went unanswered"), "{both}");
 
         // A later capture answers one of them.
@@ -9211,7 +8821,7 @@ mod tests {
             .await
             .unwrap();
 
-        let one = get_body(&app, &cookie, "/ui/ops").await;
+        let one = get_body(&app, &cookie, "/ui/insights").await;
         assert!(one.contains("1 went unanswered"), "{one}");
         assert!(one.contains("is\n  <a href=\"/ui/capture#gaps\""), "{one}");
     }
@@ -9688,7 +9298,7 @@ mod tests {
     #[tokio::test]
     async fn the_nav_is_the_same_width_on_every_page() {
         let (app, cookie) = app_with_session().await;
-        for uri in ["/ui/capture", "/ui/search", "/ui/ops"] {
+        for uri in ["/ui/capture", "/ui/search", "/ui/insights"] {
             let page = get_body(&app, &cookie, uri).await;
             let bar = page.find(r#"class="topbar""#).expect("a top bar");
             let shell = page.find(r#"class="shell"#).expect("a shell");
@@ -9715,7 +9325,7 @@ mod tests {
             "search is a three-region page: {search}"
         );
 
-        let ops = get_body(&app, &cookie, "/ui/ops").await;
+        let ops = get_body(&app, &cookie, "/ui/insights").await;
         assert!(
             ops.contains("regions-table"),
             "housekeeping is a table and has no reading measure: {ops}"
@@ -9771,7 +9381,7 @@ mod tests {
         // No page says how wide it is any more.
         for (uri, body) in [
             ("/ui/search", &search),
-            ("/ui/ops", &ops),
+            ("/ui/insights", &ops),
             ("/ui/capture", &capture),
             ("/ui/ask", &ask),
         ] {
@@ -10803,7 +10413,7 @@ mod tests {
         );
         assert!(detail.contains("why was this asked"), "{detail}");
 
-        let ops = get_body(&app, &cookie, "/ui/ops").await;
+        let ops = get_body(&app, &cookie, "/ui/insights").await;
         assert!(ops.contains("Generated"), "{ops}");
         assert!(
             ops.contains(&format!("/ui/ops/artifacts/{}/deprecate", g.id)),
@@ -10823,7 +10433,7 @@ mod tests {
     #[tokio::test]
     async fn the_pursuit_section_is_not_there_when_pursuits_are_off() {
         let (app, cookie) = app_with_session().await;
-        let ops = get_body(&app, &cookie, "/ui/ops").await;
+        let ops = get_body(&app, &cookie, "/ui/insights").await;
         assert!(!ops.contains("<h3>Pursuits</h3>"), "{ops}");
     }
